+++ duckscript_sdk/src/sdk/std/var/mod.rs
@@ -1,3 +1,4 @@
+mod get_all_var_names;
 mod get_by_name;
 pub(crate) mod set;
 mod set_by_name;
@@ -11,6 +12,7 @@ static PACKAGE: &str = "var";
 pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptError> {
     let package = pckg::concat(parent, PACKAGE);
 
+    commands.set(get_all_var_names::create(&package))?;
     commands.set(get_by_name::create(&package))?;
     commands.set(set::create(&package))?;
     commands.set(set_by_name::create(&package))?;